refactor: create src folder and move components, redux, prisma, and styles dirs into it
