+++ pages/_app.tsx
@@ -1,8 +1,8 @@
 import { Provider as ReduxProvider}  from 'react-redux'
 import { Provider as NextAuthProvider} from 'next-auth/client'
 import { AppProps } from 'next/app'
-import '../styles/globals.css'
-import store from '../redux/store'
+import '../src/styles/globals.css'
+import store from '../src/redux/store'
 
 
 export default function App({ Component, pageProps }: AppProps) {
@@ -14,3 +14,4 @@ export default function App({ Component, pageProps }: AppProps) {
     </ReduxProvider>
   )
 }
+
